Refactor debug canvas and move it to excalidraw-app
