+++ packages/excalidraw/components/App.tsx
@@ -435,7 +435,6 @@ import { actionTextAutoResize } from "../actions/actionTextAutoResize";
 import { getVisibleSceneBounds } from "../element/bounds";
 import { isMaybeMermaidDefinition } from "../mermaid";
 import { mutateElbowArrow } from "../element/routing";
-import DebugCanvas from "./canvases/DebugCanvas";
 import {
   FlowChartCreator,
   FlowChartNavigator,
@@ -1694,13 +1693,6 @@ class App extends React.Component<AppProps, AppState> {
                               this.flowChartCreator.pendingNodes,
                           }}
                         />
-                        {(import.meta.env.MODE === ENV.TEST ||
-                          import.meta.env.DEV) && (
-                          <DebugCanvas
-                            appState={this.state}
-                            scale={window.devicePixelRatio}
-                          />
-                        )}
                         <InteractiveCanvas
                           containerRef={this.excalidrawContainerRef}
                           canvas={this.interactiveCanvas}
